test: rename + update of text_sampling helpers
